pyproject.toml
[project]
name = "mongodb-childcare"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=33.1.0",
    "jupyter>=1.1.1",
    "pymongo>=4.10.1",
    "python-dotenv>=1.0.1",
]
